Code restructuring and added user related code
only register in backend
