+++ server/package.json
@@ -2,22 +2,25 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "src/index.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon app.js"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "bcryptjs": "^2.4.3",
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.18.3",
     "express-fileupload": "^1.4.3",
+    "express-session": "^1.18.0",
     "imgur": "^2.4.2",
     "mongodb": "^6.4.0",
     "mongoose": "^8.2.1",
     "multer": "^1.4.5-lts.1",
+    "passport": "^0.7.0",
     "pg-promise": "^11.5.4"
   }
 }
