+++ server/src/index.js
@@ -1,0 +1,8 @@
+import app from './app.js'
+import connectDB from './database.js'
+
+connectDB();
+
+app.listen(app.get('port'), () => {
+    console.log('Server running on port:', app.get('port'))
+})
